requirements.txt
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-dysmsapi==1.0.0
backcall==0.1.0
colorama==0.4.1
decorator==4.4.0
Django==2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.9.3
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jieba==0.39
jmespath==0.9.4
mysqlclient==1.4.2.post1
parso==0.3.4
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
Pygments==2.3.1
python-memcached==1.59
pytz==2018.9
shortuuid==0.5.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
Whoosh==2.7.4
